COMMERCE-5049 commerce-machine-learning-api Apply new naming pattern to commerce machine learning
